requirements.txt
PyMeta3==0.5.1
pystache==0.5.4
PyYAML==5.3
